Fix resource name verify for prefix match
verifyUIDOrName()'s uses a client Find() functions which uses filter
to find a given name or id. Because it's a filter, it performs a
prefix match. This results in false positive results.

For example, if there's an existing resource "worker2", and a new
resource "worker" is requested, due to the filter's prefix match,
it results in ambiguous error - already exists.

This change adds an extra check of the filtered object by matching
it with the given name/id.